requirements.txt
cython
scipy
shapely
timm
h5py
wandb
fire
opencv-python
pandas
braceexpand
torch-ema
torchmetrics==0.11.4
setuptools==59.5.0
webdataset>=0.2.5
numpy==1.23.0